Move types to feature-specific folder
