+++ frontend/src/features/noteTaking/notebooks/types/notebookStore.types.ts
