+++ frontend/src/features/noteTaking/notes/types/notesStore.types.ts
